+++ demo/children.js
@@ -1,3 +1,5 @@
+// NOTE: children game obj is not usable yet, will come in next major release!
+
 kaboom()
 
 loadSprite("bean", "/sprites/bean.png")
@@ -20,13 +22,13 @@ for (let i = 12; i < 24; i++) {
       speed: i * 8,
     },
   ])
-	
+
 }
 
 nucleus.onUpdate(() => {
 
   nucleus.pos = mousePos()
-	
+
   // update children
   nucleus.children.forEach((child) => {
     child.angle += child.speed * dt()