+++ scripts/build.js
@@ -32,7 +32,6 @@ fmts.forEach((fmt) => {
 	esbuild.build({
 		bundle: true,
 		sourcemap: true,
-		target: "es6",
 		minify: !dev,
 		keepNames: true,
 		loader: {